Add tests for controllers and service

+++ src/main/java/com/mate/test/autoservice/mateautoservice/service/MasterService.java
@@ -1,11 +1,12 @@
 package com.mate.test.autoservice.mateautoservice.service;
 
 import com.mate.test.autoservice.mateautoservice.model.Master;
+import java.math.BigDecimal;
 
 public interface MasterService {
     Master save(Master master);
 
     Master getById(Long id);
 
-    Double paidForServicesForMaster(Long id);
+    BigDecimal paidForServicesForMaster(Long id);
 }
